+++ docs/about/changelog.md
@@ -1,5 +1,16 @@
 # Changelog
 
+Updates for the website.
+
+For bugs or feature requests, message me on Discord (Pleb#0025) or create an issue on [Github](https://github.com/pvme/pvme.github.io).
+
+## November, 13, 2022
+
+- Reorganized about and home page (improves navigation on mobile).
+- Better alias support
+  - Alias for forums/words/extra channels
+  - Category emojis
+
 ## November, 5, 2022
 
 - New site structure: